+++ projectq/ops/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 #   Copyright 2017 ProjectQ-Framework (www.projectq.ch)
+#   Copyright 2021 <Huawei Technologies Co., Ltd>
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -15,34 +16,22 @@
 
 """ProjectQ module containing all basic gates (operations)"""
 
-from ._basics import (
-    BasicGate,
-    BasicMathGate,
-    BasicPhaseGate,
-    BasicRotationGate,
-    ClassicalInstructionGate,
-    FastForwardingGate,
-    MatrixGate,
-    NotInvertible,
-    NotMergeable,
-    SelfInverseGate,
-)
-from ._command import Command, CtrlAll, IncompatibleControlState, apply_command
-from ._gates import *
-from ._metagates import (
-    All,
-    C,
-    ControlledGate,
-    DaggeredGate,
-    Tensor,
-    get_inverse,
-    is_identity,
-)
-from ._qaagate import QAA
-from ._qftgate import QFT, QFTGate
-from ._qpegate import QPE
-from ._qubit_operator import QubitOperator
-from ._shortcuts import *
-from ._state_prep import StatePreparation
-from ._time_evolution import TimeEvolution
-from ._uniformly_controlled_rotation import UniformlyControlledRy, UniformlyControlledRz
+import importlib
+import pkgutil
+from pathlib import Path
+
+# Allow extending this namespace.
+__path__ = pkgutil.extend_path(__path__, __name__)
+
+from ._base import *
+
+for (_, pkg_name, _) in pkgutil.iter_modules(path=__path__):
+    if pkg_name != '_base':
+        imported_module = importlib.import_module('.' + pkg_name, package=__name__)
+
+        if hasattr(imported_module, '__all__'):
+            for symbol_name in imported_module.__all__:
+                globals().setdefault(symbol_name, getattr(imported_module, symbol_name))
+        else:
+            for attr_name in dir(imported_module):
+                globals().setdefault(attr_name, getattr(imported_module, attr_name))