+++ projectq/setups/ionq.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 #   Copyright 2021 ProjectQ-Framework (www.projectq.ch)
+#   Copyright 2021 <Huawei Technologies Co., Ltd>
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -20,9 +21,9 @@ Defines a setup allowing to compile code for IonQ trapped ion devices:
 ->The 11 qubit device
 ->The 29 qubits simulator
 """
-from projectq.backends._exceptions import DeviceOfflineError
-from projectq.backends._ionq._ionq_http_client import show_devices
-from projectq.backends._ionq._ionq_mapper import BoundedQubitMapper
+from projectq.backends._base import DeviceOfflineError
+from projectq.backends._web._ionq._ionq_http_client import show_devices
+from projectq.backends._web._ionq._ionq_mapper import BoundedQubitMapper
 from projectq.ops import (
     Barrier,
     H,
@@ -67,6 +68,3 @@ def get_engine_list(token=None, device=None):
         other_gates=(Barrier,),
     )
     return engine_list + [mapper]
-
-
-__all__ = ['get_engine_list']