+++ projectq/types/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 #   Copyright 2017 ProjectQ-Framework (www.projectq.ch)
+#   Copyright 2021 <Huawei Technologies Co., Ltd>
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -15,4 +16,5 @@
 
 """ProjectQ module containing all basic types"""
 
+from ._enum import CtrlAll
 from ._qubit import BasicQubit, Qubit, Qureg, WeakQubitRef